Fixed gcc warnings in the following locations:
* libsvn_subr/io.c (svn_io_open_unique_file)
* libsvn_ra_dav/commit.c (create_activity)
* libsvn_client/commit.c (svn_client_commit)


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@838902 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_subr/io.c
@@ -90,9 +90,10 @@ svn_io_open_unique_file (apr_file_t **f,
      that the iterating portion changes instead.  Taking the pointer
      as an unsigned short int has more or less this effect. */
   int random_portion_width;
-  char *random_portion = apr_psprintf (pool, "%hu%n",
-                                       unique_name,
-                                       &random_portion_width);
+  char *random_portion = apr_psprintf 
+    (pool, "%hu%n",
+     (unsigned int)unique_name,
+     &random_portion_width);
 
   *unique_name = svn_string_dup (path, pool);
